+++ demo/constants/Colors.js
@@ -11,4 +11,6 @@ export default {
   warningText: "#666804",
   noticeBackground: tintColor,
   noticeText: "#fff",
+  backgroundColor: "#fff",
+  textColor: "gray",
 };
