+++ demo/package.json
@@ -13,10 +13,10 @@
     "lint": "npx prettier --write '{*.js,**/*.js,*.jsx,**/*.jsx}' && npx eslint '{*.js,**/*.js,*.jsx,**/*.jsx}'"
   },
   "dependencies": {
-    "ethers": "^4.0.14",
     "expo": "^31.0.3",
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-31.0.0.tar.gz",
+    "react-native-responsive-dimensions": "^2.0.1",
     "react-navigation": "^3.0.5",
     "tasit-sdk": "0.0.2"
   },
