+++ demo/screens/EthereumQuestion.js
@@ -1,12 +1,13 @@
 import React from "react";
-import { Image, Button, StyleSheet, View, Text } from "react-native";
-import { createStackNavigator, createAppContainer } from "react-navigation";
+import { Button, StyleSheet, View, Text } from "react-native";
+import { responsiveHeight } from "react-native-responsive-dimensions";
+import Colors from "../constants/Colors";
 
 export default class EthereumQuestion extends React.Component {
   render() {
     return (
       <View style={styles.container}>
-        <Text style={styles.questionText}>{`Are you new to Ethereum?`}</Text>
+        <Text style={styles.questionText}>Are you new to Ethereum?</Text>
         <View style={styles.buttonView}>
           <Button
             title="Yep"
@@ -27,16 +28,16 @@ export default class EthereumQuestion extends React.Component {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: "#fff",
+    backgroundColor: Colors.backgroundColor,
     alignItems: "center",
     justifyContent: "center",
   },
   questionText: {
-    padding: 10,
+    padding: responsiveHeight(2),
     fontSize: 30,
     textAlign: "center",
     fontWeight: "bold",
-    color: "gray",
+    color: Colors.textColor,
   },
   buttonView: {
     flexDirection: "row",
