+++ demo/screens/OnboardingHome.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Image, Button, StyleSheet, View, Text } from "react-native";
-import { createStackNavigator, createAppContainer } from "react-navigation";
+import Colors from "../constants/Colors";
 
 export default class OnboardingHome extends React.Component {
   render() {
@@ -24,7 +24,7 @@ export default class OnboardingHome extends React.Component {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: "#fff",
+    backgroundColor: Colors.backgroundColor,
     alignItems: "center",
     justifyContent: "center",
   },
@@ -32,7 +32,7 @@ const styles = StyleSheet.create({
     fontSize: 20,
     textAlign: "center",
     fontWeight: "bold",
-    color: "gray",
+    color: Colors.textColor,
   },
   buttonView: {
     flexDirection: "row",
